requirements.txt
streamlit
faiss-cpu
docx
numpy
langchain
PyPDF2
google-generativeai
python-dotenv
